requirements/local.txt
-r production.txt

django-debug-toolbar==6.0.0
